fix(sync): handle xpointer in a different section, fix styling
