+++ .prettierrc.json
@@ -5,5 +5,6 @@
   "tabWidth": 2,
   "singleQuote": true,
   "jsxSingleQuote": true,
+  "endOfLine": "lf",
   "plugins": ["prettier-plugin-tailwindcss"]
 }
