+++ apps/readest-app/src/types/kosync.ts
